# src/SvarOMatic.Backend/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Swagger - Svar-o-Matic",
    "description" : "This is the backend for a solution that allows employees of Miles AS to efficiently match incoming project requests with existing consultant expertise.\n\nSome useful links:\n- [The Svar-o-Matic repository](https://github.com/miles-no/svar-o-matic)",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "drift@miles.no"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.11"
  },
...
  "servers" : [ {
    "url" : "http://localhost:5090"
  } ],
...
  "paths" : {
    "/cv/{consultantId}/add" : {
      "post" : {
        "tags" : [ "cv" ],
        "summary" : "Add a new CV to a consultant profile",
        "description" : "Add a new CV to a consultant profile",
        "operationId" : "addCV",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "Add a new CV for a consultant in the app",
          "content" : {
            "application/octet-stream" : {
              "schema" : {
                "type" : "string",
                "format" : "binary"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_docs_auth" : [ "write:docs", "read:docs" ]
        } ]
      }
    },
...
    "/inquiry/{consultantId}/add" : {
      "post" : {
        "tags" : [ "inquiry" ],
        "summary" : "Create a new forespørsel for a consultant",
        "description" : "Create a new forespørsel for a consultant",
        "operationId" : "createInquiry",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "Data about an inquiry",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InquiryRequest"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/InquiryRequest"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/InquiryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_inquiry_auth" : [ "write:inquiry", "read:inquiry" ]
        } ]
      }
    },
    "/inquiry/{consultantId}/{inquiryId}/generateAnswers" : {
      "post" : {
        "tags" : [ "inquiry" ],
        "summary" : "Generate answers for a forespørsel",
        "description" : "Generate answers for a forespørsel",
        "operationId" : "generateAnswersForInquiry",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the request is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "inquiryId",
          "in" : "path",
          "description" : "Id of the forespørsel that the request is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_inquiry_auth" : [ "write:inquiry", "read:inquiry" ]
        } ]
      }
    },
...
    "/docsForConsultant/{consultantId}/get" : {
      "get" : {
        "tags" : [ "cv", "previous-responses", "style-guide", "consultant" ],
        "summary" : "Finds all CVs, skjemaer and style-guides applicable for a consultant",
        "description" : "Multiple docs (CV, skjema and/or style guide) values can be provided with comma separated strings",
        "operationId" : "findDocsForConsultant",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the search is limited to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "anyOf" : [ {
                      "$ref" : "#/components/schemas/StyleGuide"
                    }, {
                      "$ref" : "#/components/schemas/CV"
                    }, {
                      "$ref" : "#/components/schemas/PreviousResponse"
                    } ]
                  },
                  "x-content-type" : "application/json"
                }
              },
              "application/xml" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "anyOf" : [ {
                      "$ref" : "#/components/schemas/StyleGuide"
                    }, {
                      "$ref" : "#/components/schemas/CV"
                    }, {
                      "$ref" : "#/components/schemas/PreviousResponse"
                    } ]
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid status value"
          }
        },
        "security" : [ {
          "svaromatic_docs_auth" : [ "write:docs", "read:docs" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "StyleGuide" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string",
            "format" : "binary"
          },
          "modified" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isInUse" : {
            "type" : "boolean"
          }
        }
      },
      "CV" : {
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "content" : {
            "type" : "string",
            "format" : "binary"
          },
          "owner" : {
            "type" : "integer",
            "format" : "int64"
          },
          "modified" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PreviousResponse" : {
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "content" : {
            "type" : "string",
            "format" : "binary"
          },
          "owner" : {
            "type" : "integer",
            "format" : "int64"
          },
          "originalSubmitter" : {
            "type" : "integer",
            "format" : "int64"
          },
          "submittedOn" : {
            "type" : "string",
            "format" : "date-time"
          },
          "submittedToKunde" : {
            "type" : "string"
          },
          "ledToInterview" : {
            "type" : "boolean"
          },
          "ledToOppdrag" : {
            "type" : "boolean"
          },
          "isInUse" : {
            "type" : "boolean"
          }
        }
      },
...
      "InquiryRequest" : {
        "type" : "object",
        "properties" : {
          "kunde" : {
            "type" : "string"
          },
          "recievedOn" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "xml" : {
          "name" : "##default"
        }
      },
...
      "ApiResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "code" : 0,
          "type" : "type",
          "message" : "message"
        },
        "xml" : {
          "name" : "##default"
        }
      }
    },
...
    "securitySchemes" : {
      "svaromatic_docs_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://coming-soon.swagger.io/oauth/authorize",
            "scopes" : {
              "write:cvs" : "modify docs for a consultant",
              "read:cvs" : "read your docs"
            }
          }
        }
      },
      "svaromatic_inquiry_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://coming-soon.swagger.io/oauth/authorize",
            "scopes" : {
              "write:cvs" : "modify inquiries for a consultant",
              "read:cvs" : "read inquiries"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}